+++ CHANGELOG.md
@@ -19,6 +19,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### added
 * added `select_related` in intospect view for better query performance
 
+### Fixed
+* #847: Fix inappropriate message when response from authentication server is not OK.
 
 ## [1.3.2] 2020-03-24